Fix #71 by properly resetting fields
When the mode was changed from **authentication** to **registration** or
the other way around, the model information is reset (because of the
usage of top-level navigation), but the input fields remained on the
screen and their contents were not cleared.

Therefore, when switching between the modes, it would look like the
input fields were still filled when in fact, they were not.

The fix consists of reflecting the current model state in the two
fields.